requirements.txt
# API and Web Framework
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7

# UI
gradio==3.32.0
plotly==5.14.1
matplotlib==3.7.1
pillow==9.5.0

# Data Processing
pandas==2.0.0
numpy==1.24.3
scipy==1.10.1

# Storage
redis==4.5.4
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
alembic==1.11.1

# Networking
requests==2.28.2
discord.py==2.2.3

# Trading
lumibot==2.5.0
alpaca-trade-api==2.3.0
polygon-api-client==1.12.3
# Use a specific yfinance version compatible with lumibot
yfinance==0.2.10
backtrader==1.9.78.123

# Visualization
seaborn==0.12.2

# Testing
pytest==7.3.1

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
schedule==1.2.0
boto3==1.26.0
ujson==5.7.0

# Monitoring
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation==0.39b0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-instrumentation-redis==0.39b0
opentelemetry-instrumentation-sqlalchemy==0.39b0
opentelemetry-instrumentation-requests==0.39b0
